Create zip for certificates
